requirements.txt
Flask
fpdf2
Flask-SQLAlchemy
python-dateutil
gunicorn
openpyxl
pdfplumber
anthropic
python-dotenv
